% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyst_to_BayFolders.R
\name{Analyst_to_BayFolders}
\alias{Analyst_to_BayFolders}
\title{analyst_to_BayFolders}
\usage{
Analyst_to_BayFolders(
  data,
  folder.name = "BayLum_files",
  inflatePercent = 0.025
)
}
\arguments{
\item{data}{Data.frame containing information about each disc and grain to be used in BayLum. This function is optimized for use with the output of Analyst. Importantly, the data.frame must contain columns: "Labcode", "DoseEnv", "DoseEnv_error", "DoseSource", "DoseSource_error", "NbOfLastCycleToRemove", "Filename", "Disc#", "Grain#", "Sig1", "Sig2", "BG1", "BG2".}

\item{folder.name}{Specify a name for a folder to be created in the current working directory which will contain the output of this function (defaults to "BayLum_files").}

\item{inflatePercent}{Specify measurement uncertainty that should be incorporated (defaults to 0.025).}
}
\value{
Returns a list of vectors containing the required input for BayLum functions Generate_DataFile() and AgeS_Computation().
}
\description{
Writes files necessary to use Generate_DataFile() function of the BayLum R-package.
}
