% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit.R, R/jap_primary.R
\name{crit}
\alias{crit}
\title{crit}
\usage{
crit(criteria, error, signif = 4, limits, limittype = "<>")

crit(criteria, error, signif = 4, limits, limittype = "<>")
}
\description{
Does things.
And returns things.
}
