% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_criteria_function2.R
\name{my_criteria_function2}
\alias{my_criteria_function2}
\title{my_criteria_function2}
\usage{
my_criteria_function2(data = NULL, index, blanks = NULL, keep.alternatives = F)
}
\description{
my_criteria_function2
}
