% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_signal_function.R
\name{my_signal_function}
\alias{my_signal_function}
\title{my_signal_function}
\usage{
my_signal_function(
  Risoe.object,
  LTYPE = "OSL",
  ID = NULL,
  set = NULL,
  run = NULL,
  position = NULL,
  S.interval = 6:10,
  BG.interval = 346:395
)
}
\description{
my_signal_function
}
