% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_primary.R
\name{my_write_BayLum_files}
\alias{my_write_BayLum_files}
\title{My write BayLum files}
\usage{
my_write_BayLum_files(
  path,
  subsample.folder.name,
  Disc = NULL,
  DiscPos = NULL,
  DRenv,
  DRenv.error,
  DRsource,
  DRsource.error,
  signal.integral,
  background.integral,
  inflatePercent,
  nbOfLastCycleToRemove
)
}
\description{
Does things
And returns things
}
