% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_primary.R
\name{cam_per_d0}
\alias{cam_per_d0}
\title{Cam per D0}
\usage{
cam_per_d0(data, minD0, method = "CAM", error.mode = "propagate", ...)
}
\arguments{
\item{data}{3-column data frame containing (in order) De, De.error and D0 features. Alternatively you can specify colnames or column numbers of said features (see De/De.error/D0).}

\item{minD0}{Vector of D0 thresholds.}

\item{method}{Method of estimating central tendency in equivalent dose analysis. Currently supported are "CAM" (default) and "ADM".}

\item{...}{Arguments passed to CAM or ADM. For example one might be interested in specifying "log" or "sigmab" in the CAM procedure or "sigma_m" in the ADM procedure.}

\item{De/De.error/D0}{Column names of relevant variables in the OSL dataset. Selects the first three columns by default.}
}
\value{
Returns a dataframe of equivalent doses and D0 thresholds along with intersect coordinate(s) and error estimates.
}
\description{
Calculates equivalent dose for specified D0 thresholds.
Estimates are then plotted against chosen D0 thresholds and the point of intersect is estimated using linear regression between the two points containing the intersect.
}
