% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_GEN1.R
\name{cam_per_d0}
\alias{cam_per_d0}
\title{Cam per D0}
\usage{
cam_per_d0(OSLdata, minD0, method = "CAM", De = 1, De.error = 2, D0 = 3, ...)
}
\arguments{
\item{OSLdata}{3-column data frame containing (in order) De, De.error and D0 features. Alternatively you can specify colname or location of said features (see De/De.error/D0).}

\item{minD0}{Vector of D0 thresholds.}

\item{method}{Method of estimating central tendency in equivalent dose analysis. Currently supported are "CAM" (default) and "ADM".}

\item{sigma_m}{Intrinsic level of overdispersion in sample to use in CAM analysis.}

\item{De/De.error/D0}{Column names of relevant variables in the OSL dataset. Selects the first three columns by default.}
}
\value{
Returns a dataframe of calculated equivalent doses and corresponding D0 thresholds. Also return coordinate(s) of intersect(s) along with estimated error. To see how error is estimated, see "details".
}
\description{
Calculates equivalent dose for each specified D0 threshold.
Estimates are then plotted against D0 thresholds and the point of intersect is estimated using linear regression between the two points containing the intersect.
}
\details{
$y_1+y_2$
}
