% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_primary.R
\name{my_criteria_function2}
\alias{my_criteria_function2}
\title{My criteria function}
\usage{
my_criteria_function2(data = NULL, index, blanks = NULL, keep.alternatives = F)
}
\description{
Does things.
And returns things.
}
