% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_primary.R, R/my_readerrate.R
\name{my_readerrate}
\alias{my_readerrate}
\title{My readerrate retriever}
\usage{
my_readerrate(
  dr.model,
  date,
  method = "average",
  day.lim = 60,
  data.path =
    "C:/Users/fhaba/OneDrive - Danmarks Tekniske Universitet/OSLdata/Calibration/calibration.csv"
)

my_readerrate(
  dr.model,
  date,
  method = "average",
  day.lim = 60,
  data.path =
    "C:/Users/fhaba/OneDrive - Danmarks Tekniske Universitet/OSLdata/Calibration/calibration.csv"
)
}
\description{
Does things.
And returns things.
}
