% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_primary.R, R/my_signal_function.R
\name{my_signal_function}
\alias{my_signal_function}
\title{My Signal Function}
\usage{
my_signal_function(
  Risoe.object,
  LTYPE = "OSL",
  ID = NULL,
  set = NULL,
  run = NULL,
  position = NULL,
  S.interval = 6:10,
  BG.interval = 346:395
)

my_signal_function(
  Risoe.object,
  LTYPE = "OSL",
  ID = NULL,
  set = NULL,
  run = NULL,
  position = NULL,
  S.interval = 6:10,
  BG.interval = 346:395
)
}
\description{
Does things.
And returns things.
}
