% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jap_primary.R
\name{my_write_BayLum_files}
\alias{my_write_BayLum_files}
\title{My write BayLum files}
\usage{
my_write_BayLum_files(
  path,
  subsample.folder.names,
  Disc = NULL,
  DiscPos = NULL,
  DRenv,
  DRenv.error,
  DRsource,
  DRsource.error,
  begin.signal.integral,
  end.signal.integral,
  begin.background.integral,
  end.background.integral,
  inflatePercent = 0.025,
  nbOfLastCycleToRemove = 2
)
}
\description{
Does things
And returns things
}
